requirements.txt
requests>=2.25.0
colorama>=0.4.4
holidays==0.13
bs4>=0.0.1
beautifulsoup4>=4.9.3
html5lib>=1.1
lxml>=4.6.3

# Development dependencies
pytest>=6.0.0
pytest-cov>=2.12.0



# Backend dependencies for chatbot.py
fastapi==0.115.12
uvicorn==0.34.1
python-dotenv==1.1.0
langchain-openai==0.3.12
langchain-community==0.3.21
langchain-chroma==0.2.3
chromadb==0.6.3
huggingface-hub==0.30.2
starlette==0.46.2
pydantic==2.11.3
typing_extensions==4.13.2
sentence-transformers
langchain-huggingface
